requirements.txt
discord.py==1.3.4
python-dotenv
flake8
pymongo
dnspython